refactor(autumn-js): delete unread .d.mts twins instead of pointering them
package.json resolves every subpath through a single "types" key pointing at
the .d.ts, so the .d.mts files tsup emits beside them are never read by any
consumer. Deleting them replaces the twin-matching, default-export forwarding
and orphan-chunk reference counting with one unlink pass.

Renamed to link-sdk-types.mjs: repointing the workspace SDK import is the
pass that matters, and "dedupe" described the part that just went away.

Package is 304 files / 5,347,678 bytes, down from 312 / 5,347,911. Verified
identical to the previous build: every subpath resolves under both require
and import, a consumer type-checks with 0 errors under bundler resolution and
the same 1 pre-existing better-auth error under node16, the attw table is
unchanged across all 10 subpaths, and no @useautumn/sdk specifier survives in
any shipped declaration.
